+++ amd/comgr/src/comgr-compiler.h
@@ -150,6 +150,7 @@ public:
   amd_comgr_status_t linkToRelocatable();
   amd_comgr_status_t linkToExecutable();
   amd_comgr_status_t compileToFatBin();
+  amd_comgr_status_t compileToExecutable();
 
   amd_comgr_language_t getLanguage() const { return ActionInfo->Language; }
 };