+++ amd/comgr/docs/ReleaseNotes.md
@@ -100,6 +100,14 @@ of the flag handling to toolchain. Currently only supports HIP.
 - (Data Type) AMD\_COMGR\_DATA\_KIND\_AR\_BUNDLE
   - These data kinds can now be passed to an AMD\_COMGR\_ACTION\_LINK\_BC\_TO\_BC
 action, and Comgr will internally unbundle and link via the OffloadBundler and linkInModule APIs.
+- (Language Type) AMD\_COMGR\_LANGUAGE\_LLVM\_IR
+  - This language can now be passed to AMD\_COMGR\_ACTION\_COMPILE\_\* actions
+  to enable compilation of LLVM IR (.ll or .bc) files. This is useful for MLIR
+  contexts.
+- (Action) AMD\_COMGR\_ACTION\_COMPILE\_SOURCE\_TO\_EXECUTABLE
+  - This action allows compilation from source directly to executable, including
+  linking device libraries.
+
 
 Deprecated Comgr Actions and Data Types
 ---------------------------------------